refactor: 💡 Convert desktop tests to mjs and use es6 imports
