+++ e2e-tests/desktop/fixtures/baseTest.mjs
@@ -3,10 +3,11 @@
  * SPDX-License-Identifier: BUSL-1.1
  */
 
-const { mergeTests } = require('@playwright/test');
-const { electronTest } = require('../fixtures/electronTest');
-const { authenticateTest } = require('./authenticateTest');
+import { mergeTests } from '@playwright/test';
+import { electronTest } from './electronTest.mjs';
+import { authenticateTest } from './authenticateTest.mjs';
 
 // Redundant since authenticateTest already extends electronTest but this is for demonstration purposes
 // The last item in the list will override any previous ones in a fixture name collision
-exports.test = mergeTests(authenticateTest, electronTest);
+export const test = mergeTests(authenticateTest, electronTest);
+export { expect } from '@playwright/test';
