+++ e2e-tests/desktop/tests/settings.spec.mjs
@@ -3,8 +3,7 @@
  * SPDX-License-Identifier: BUSL-1.1
  */
 
-const { test } = require('../fixtures/baseTest');
-const { expect } = require('@playwright/test');
+import { expect, test } from '../fixtures/baseTest.mjs';
 
 test.describe('Settings Page tests', async () => {
   test.beforeEach(async ({ authedPage }) => {
